requirements.txt
# Discord Finance Bot Requirements
discord.py>=2.3.0
python-dotenv>=1.0.0
gocardless-pro>=3.3.0

# Optional: for CSV processing
pandas>=1.5.0

# Optional: for Google Sheets integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
